requirements.txt
aiohttp==3.8.4 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "4.0"
arrow==1.2.3 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
binaryornot==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
chardet==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.4 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cookiecutter==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
croniter==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
cryptography==41.0.1 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
dj-stripe==2.8.1 ; python_version >= "3.10" and python_version < "4.0"
django-allauth==0.50.0 ; python_version >= "3.10" and python_version < "4.0"
django-anymail[mailgun]==9.2 ; python_version >= "3.10" and python_version < "4.0"
django-autoslug==1.9.9 ; python_version >= "3.10" and python_version < "4.0"
django-environ==0.10.0 ; python_version >= "3.10" and python_version < "4"
django-extensions==3.2.3 ; python_version >= "3.10" and python_version < "4.0"
django-filter==22.1 ; python_version >= "3.10" and python_version < "4.0"
django-model-utils==4.3.1 ; python_version >= "3.10" and python_version < "4.0"
django-ninja==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
django-picklefield==3.1 ; python_version >= "3.10" and python_version < "4"
django-q2==1.5.4 ; python_version >= "3.10" and python_version < "4"
django-widget-tweaks==1.4.12 ; python_version >= "3.10" and python_version < "4.0"
django==4.2.3 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.2 ; python_version >= "3.10" and python_version < "3.11"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.23.3 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.4.3 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
openai==0.27.8 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.6 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.11 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.15.1 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-slugify==8.0.1 ; python_version >= "3.10" and python_version < "4.0"
python-webpack-boilerplate==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
stripe==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.5.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "4.0"
